requirements.txt
-e paella
openai
anthropic
google-genai
google-generativeai
langchain-community
langchain-openai
langchain-google-genai
langchain-deepseek
langchain-ollama
tiktoken
rich
